Refactor out of MainActivity into RecalibrateFragment all the recalibration-related code
