+++ app/src/main/java/com/kamron/pogoiv/activities/MainFragment.java
@@ -6,7 +6,6 @@ import android.content.Context;
 import android.content.Intent;
 import android.content.IntentFilter;
 import android.content.pm.PackageManager;
-import android.graphics.Point;
 import android.net.Uri;
 import android.os.Bundle;
 import android.preference.PreferenceManager;
@@ -17,22 +16,16 @@ import android.support.v4.app.Fragment;
 import android.support.v4.content.LocalBroadcastManager;
 import android.support.v4.view.ViewCompat;
 import android.support.v7.app.AlertDialog;
-import android.view.Display;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.webkit.WebView;
-import android.webkit.WebViewClient;
 import android.widget.AdapterView;
 import android.widget.Button;
 import android.widget.ImageButton;
-import android.widget.LinearLayout;
 import android.widget.NumberPicker;
-import android.widget.ScrollView;
 import android.widget.Spinner;
 import android.widget.TextView;
 
-import com.kamron.pogoiv.BuildConfig;
 import com.kamron.pogoiv.GoIVSettings;
 import com.kamron.pogoiv.NpTrainerLevelPickerListener;
 import com.kamron.pogoiv.R;
@@ -49,11 +42,7 @@ public class MainFragment extends Fragment {
     private static final String EXTRA_BUTTON_TEXT_RES_ID = "btn_txt_res_id";
     private static final String EXTRA_BUTTON_ENABLED = "btn_enabled";
 
-    private static final String youtubeTutorialCalibrationUrl = "https://www.youtube.com/embed/w7dNEW1FLjQ?rel=0";
 
-
-    @BindView(R.id.mainScrollView)
-    ScrollView mainScrollView;
 
     @BindView(R.id.startButton)
     Button startButton;
@@ -67,18 +56,6 @@ public class MainFragment extends Fragment {
     @BindView(R.id.versionNumber)
     TextView versionNumber;
 
-    @BindView(R.id.optimizationWarningLayout)
-    LinearLayout optimizationWarningLayout;
-
-    @BindView(R.id.shouldRunOptimizationAgainWarning)
-    TextView shouldRunOptimizationAgainWarning;
-
-    @BindView(R.id.neverRunOptimizationWarning)
-    TextView neverRunOptimizationWarning;
-
-    @BindView(R.id.nonStandardScreenWarning)
-    TextView nonStandardScreenWarning;
-
     @BindView(R.id.githubButton)
     ImageButton githubButton;
 
@@ -87,18 +64,6 @@ public class MainFragment extends Fragment {
 
     @BindView(R.id.helpButton)
     Button helpButton;
-
-    @BindView(R.id.recalibrationHelpButton)
-    Button recalibrationHelpButton;
-
-    @BindView(R.id.recalibrationHelpButton2)
-    Button recalibrationHelpButton2;
-
-    @BindView(R.id.optimizationVideoTutorialLayout)
-    LinearLayout optimizationVideoTutorialLayout;
-
-    @BindView(R.id.optimizationVideoTutorial)
-    WebView optimizationVideoTutorial;
 
 
     private final BroadcastReceiver launchButtonChange = new BroadcastReceiver() {
@@ -168,58 +133,11 @@ public class MainFragment extends Fragment {
      */
     private void initiateGui() {
         versionNumber.setText(String.format("v%s", getVersionName()));
-        setupTutorialButton();
-        hideWebViewIfOfflineFlavour();
-        initiateOptimizationWarning();
         initiateLevelPicker();
         initiateTeamPickerSpinner();
         initiateHelpButton();
         initiateCommunityButtons();
         initiateStartButton();
-    }
-
-    private void hideWebViewIfOfflineFlavour() {
-        if (BuildConfig.FLAVOR.toLowerCase().contains("offline")) {
-            optimizationVideoTutorial.setVisibility(View.GONE);
-        }
-    }
-
-    /**
-     * Show the optimization-warning and its components depending on if the user hasn't a manual screen calibration
-     * saved, if the calibration isn't updated and if the device has weird screen ratio.
-     */
-    private void initiateOptimizationWarning() {
-        GoIVSettings settings = GoIVSettings.getInstance(getContext());
-        if (settings.hasUpToDateManualScanCalibration()) {
-            optimizationWarningLayout.setVisibility(View.GONE); // Ensure the layout isn't visible
-
-        } else {
-            optimizationWarningLayout.setVisibility(View.VISIBLE);
-
-            if (settings.hasManualScanCalibration()) {
-                // Has outdated calibration
-                shouldRunOptimizationAgainWarning.setVisibility(View.VISIBLE);
-
-            } else {
-                // Has never calibrated
-                neverRunOptimizationWarning.setVisibility(View.VISIBLE);
-
-                Activity activity = getActivity();
-                if (activity == null) {
-                    return;
-                }
-                // If the screen ratio isn't standard the user must run calibration
-                Display display = activity.getWindowManager().getDefaultDisplay();
-                Point size = new Point();
-                display.getRealSize(size);
-                float ratio = (float) size.x / size.y;
-                float standardRatio = 9 / 16f;
-                float tolerance = 1 / 400f;
-                if (ratio < (standardRatio - tolerance) || ratio > (standardRatio + tolerance)) {
-                    nonStandardScreenWarning.setVisibility(View.VISIBLE);
-                }
-            }
-        }
     }
 
     /**
@@ -315,57 +233,6 @@ public class MainFragment extends Fragment {
         NpTrainerLevelPickerListener listener = new NpTrainerLevelPickerListener(getContext());
         trainerLevelPicker.setOnScrollListener(listener);
         trainerLevelPicker.setOnValueChangedListener(listener);
-    }
-
-    /**
-     * Makes the help-buttons load and navigate to the tutorial youtube webview, or open the browser if using offline
-     * build.
-     */
-    private void setupTutorialButton() {
-        View.OnClickListener tutorialListener = new View.OnClickListener() {
-            @Override public void onClick(View v) {
-                if (BuildConfig.FLAVOR.toLowerCase().contains("online")) {
-
-                    if (optimizationVideoTutorialLayout.getVisibility() == View.GONE) {
-                        optimizationVideoTutorialLayout.setVisibility(View.VISIBLE);
-
-                        String frameVideo = "<html><iframe width=\"310\" height=\"480\" src=\""
-                                + youtubeTutorialCalibrationUrl
-                                + "\" frameborder=\"0\" gesture=\"media\" allow=\"encrypted-media\" "
-                                + "allowfullscreen></iframe></html>";
-
-                        optimizationVideoTutorial.setWebViewClient(new WebViewClient() {
-                            @Override
-                            public boolean shouldOverrideUrlLoading(WebView view, String url) {
-                                return false;
-                            }
-
-                        });
-                        optimizationVideoTutorial.getSettings().setJavaScriptEnabled(true);
-                        optimizationVideoTutorial.loadData(frameVideo, "text/html", "utf-8");
-
-                        mainScrollView.post(new Runnable() {
-                            @Override
-                            public void run() {
-                                mainScrollView.smoothScrollTo(0, optimizationVideoTutorial.getTop());
-                            }
-                        });
-                    } else {
-                        optimizationVideoTutorial.stopLoading();
-                        optimizationVideoTutorialLayout.setVisibility(View.GONE);
-                    }
-
-                } else {
-                    // Running offline version, we cant load the webpage inserted into the app, we need to open browser.
-                    Intent i = new Intent(Intent.ACTION_VIEW);
-                    i.setData(Uri.parse(youtubeTutorialCalibrationUrl));
-                    startActivity(i);
-                }
-            }
-        };
-
-        recalibrationHelpButton.setOnClickListener(tutorialListener);
-        recalibrationHelpButton2.setOnClickListener(tutorialListener);
     }
 
     private String getVersionName() {
